added check purge date range

+++ araqne-logdb/src/main/java/org/araqne/logdb/msgbus/ManagementPlugin.java
@@ -153,7 +153,7 @@ public class ManagementPlugin {
 					errorParams.put("remote_ip", getRemoteAddr(session));
 					throw new MsgbusException("logdb", "invalid password", errorParams);
 				}
-				
+
 				throw e;
 			}
 		}
@@ -166,7 +166,7 @@ public class ManagementPlugin {
 
 		session.setProperty("araqne_logdb_session", dbSession);
 	}
-	
+
 	private String getRemoteAddr(Session session) {
 		return session.getRemoteAddress() != null ? session.getRemoteAddress().getHostAddress() : null;
 	}
@@ -720,6 +720,9 @@ public class ManagementPlugin {
 		} catch (ParseException e) {
 			throw new MsgbusException("logdb", "not-parse-date");
 		}
+
+		if ((fromDay != null && toDay != null) && fromDay.after(toDay))
+			throw new IllegalArgumentException("araqne-logdb: invalid date range(from: " + fromDay + ", to: " + toDay);
 
 		storage.purge(tableName, fromDay, toDay);
